15.12

+++ application/views/consent/ev_quota/v_add_quota.php
@@ -178,7 +178,7 @@ function check_quota() {
         check = document.getElementById("quota" + i).value;
 
         if (check != "") {
-            value_quota += parseInt(check);
+            value_quota += parseFloat(check);
 
         }
         // if
